requirements.txt
streamlit
openai
python-dotenv
Pillow
pandas
plotly
numpy